Implemented cachematrix
The makeCacheMatrix(x) function creates a list with getters and setters
function to access the provided matrix (x) and its inverse. When
setting a new matrix, its inverse is set to NULL.
The cacheSolve(x) returns the inverse of the matrix stored in x (where
x is an object created by makeCacheMatrix). This function verifies if
the value stored as the inverse (x$getinverse()) is not NULL, in which
case it just returns the cached value x$getinverse(). If, on the other
hand, x$getinverse() is NULL, than the inverse is calculated with
solve(), stored with x$setinverse() and return to the user. Whern the
returned value comes from the cache, the message 'getting cached data'
is displayed.